feat: 팟캐스트 조회 기능 구현
S13P21D101-91

+++ backend/my-issue/src/main/java/com/ssafy/myissue/podcast/domain/PodcastSubtitle.java
@@ -25,7 +25,7 @@ public class PodcastSubtitle {
     private String line;
 
     @Column(name = "start_time")
-    private int startTime;
+    private double startTime;
 
     public static PodcastSubtitle of(Podcast podcast, int speaker, String line, double startTime) {
         return PodcastSubtitle.builder()
